Move some classes out
